chore: refactor, add benches, add readem
