+++ vectorstore_memory_test.go
